[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -2,14 +2,14 @@
 
 using namespace std;
 
-int Hyo(int a, int b) {
-    int sum;
+long Hyo(long a, long b) {
+    long sum;
     sum = (a + b) * (a - b);
     return sum;
 }
 
 int main() {
-	int a, b;
+	long a, b;
     
     cin >> a >> b;
